Improve stability of FPS calibration.
We would give the system a few frames to stabilize after a count update, but this
logic was based on frameSwapped which kept firing on the render thread, so
we did in fact not wait at all if an animator was running or something else
was triggering updates on the render thread.

Switch to using beforeSynchronizing instead which won't fire while the GUI is
busy.