Also accept results for > 60 fps
On machines where the frame rate is not capped at 60 fps,
we could easily get into the state where the test would
detect failure even if the frame rate was higher than 60.
We would suddenly end up with 0 ops/frame for these cases.
If we can do X ops in 62 fps, then we can probably also do
it in 60, so we just accept these. This seems to stabilize
the results to the point where they are actually
comparable.