Apply a different strategy for estimating averages
Rather than running a fixed amount of samples and discarding
the extremes, sort results into buckets which fall within
5% of each other. As soon as a bucket has X samples, use that
for average. This helps us to root out occational system
load which tends to clutter the results with low values.

In addition, include standard deviation, error, coefficient
of variance, so that we can show with what certainty the
numbers were acquired.