Long live the frame-count shell!
The default shell behavior of trying to find stable FPS that a system
can manage with a given item count is useful, but a little flawed for
automated testing purposes. Many systems either do not have (or cannot
keep) a stable framerate with swap, which makes results somewhat
variant.

frame-count takes a different approach: it attempts to disable the swap
delay altogether, turning the benchmark then into counting frames a
system is able to create a set number of items for.

For instance:
    ./qmlbench --shell frame-count benchmark/creation/delegates_text.qml --count=1000

... will attempt to run the delegates_text benchmark with 1000 text
items, and count how many frames the system is able to produce in a
given time frame.

Note that at least on Mac, a swap interval of 0 appears to not be
respected (due to Beam Sync), so you need to set a quite high item count
in order for this to give meaningful results, rather than just counting
a total number of frames.

In general, you likely want a quite high count parameter anyway, such
that there will be stress to reach it, but this parameter will be
somewhat system-dependent (as a tiny little CPU isn't going to have
such an easy time making it as a beefy desktop).

Future potential improvements:
* Specify the walltime to run for (also probably useful for the static shell)
* Figure out how to handle default counts properly, rather than requiring
  that they be specified (perhaps also useful for the static shell)?

Change-Id: I9dae7babce5f0518269a7c5dd68793d0bcbd1de9
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>