Static shells: Use the benchmark count if one was not explicitly provided
This provides at least some kind of plug-and-play. A followup change
will attempt to use a sensible count for each benchmark.

At the same time, ensure that we pass 'count' to the benchmark on
creation to avoid wasteful item creations that aren't necessary as part
of the benchmark (the default value of 'count' specified in the
benchmark).

Change-Id: I0eebbeb2de6b777b74317e2171041915ec41700f
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>