Add experimental support for a staticCount value in shells
The idea here is that shells can set a static item count that they feel
will be difficult to reach. This value is taken and multiplied by
--hardware-multiplier on benchmark creation to set 'count' to a value
that makes sense for that benchmark, and for that hardware.

We do not want to override the 'count' value for this (at least, not
yet), as such a change would impact the default shell, so introduce it
as a new property.

For now, we have three levels of complexity in the creation benchmarks:
5000 (for Items that don't render anything),  2500 (for the simplistic
benchmarks like Rectangle), 1000 for the text benchmarks (and benchmarks that
create text like buttons). The one exception is delegates_qobject (which
is exceptionally fast).

These may need further tweaking, but this seems to be a sensible
starting point.

Anything except creation benchmarks has _not_ been covered by this
change! For the time being, I am_only_ focusing on creation benchmarks.

Change-Id: I07931fd53139d093d356fa3d8d3e4e95242aafcd
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>