Add static count properties to other benchmarks
Some of these are "skipped" for now (gputhroughput and part of images/
being the key pieces). gputhroughput is a bit "difficult" as it's not
really a throughput benchmark so much as a hardware validation tool. The
images (sprite specifically) are a bit strange in that they seem
insanely variant, so will need specific attention to fix.

Change-Id: Ia967786361a0bf9e5c0fdd4b37ba64e7f1a9abe5
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>