Remove sprite-* benchmarks
sprite-sequence isn't really of much use. It was written to demonstrate
that SpriteSequence isn't really very performant.

sprite-image was written to compare against SpriteSequence, and while a little
more useful, the implementation is inherently flaky for two reasons:

* It uses Math.random to calculate wait times
* It uses PauseAnimation, which means that the test periodically doesn't
  run at full throttle

Since it isn't terribly useful as a benchmark for regression testing,
let's simply remove the two of them.

Change-Id: I594928c441f5e6240522881d27711e7580e86b91
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>