Alter standard deviation calculation
As with many things mathematical, there is more than one way to skin a
cat. In the case of standard deviation, there is "uncorrected sample
standard deviation" (or "population standard deviation"), and there is
"corrected sample standard deviation".

There's probably more, but the former is what we're using now, and the latter
is what this change introduces.

Quoting Erik:
    the "issue" is that the sqrt biases toward small numbers because it's a
    non-linear function. so the uncorrected one will give smaller x_i - avg
    a higher "weight" than the outliers where x_i-avg is bigger.

Note: This will cause an increase in CoV values (since they are based
off of stddev). This is probably acceptable due to the increased
"precision" gained though.

Change-Id: I0e9c8bd245de47fe7099baefc2966b744ec78028
Reported-by: Erik Verbruggen <erik.verbruggen@qt.io>
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>