"Get rid" of the changing_textcanvas benchmark
The original benchmark wasn't all that useful. It created a canvas for
every piece of text, which meant that it wasn't really a fair comparison
of methods against changing_texts.

At the same time, such a comparison isn't really useful anyway: Text is
always going to win. So, at the same time as we simplify that to make it
*just* about text rendering, move it to a new dedicated "canvas"
subfolder and rename it to make it clear that it's just benchmarking
text rendering.

In the future, we will likely want to add more canvas tests.

Change-Id: I63ff1bd53357e0ae397afe52cfb8381e5ae2456c
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>