changing_text_oneofmany: Simplify benchmark
The original purpose of this benchmark was a little muddled. It seems
like it was being used to test isolation clip (but we already have that
through changing_over_isolated{*}) and changing a single piece of text.

We can do this simpler, by stealing most of changing_texts'
implementation but changing the text binding. This way, we gain a
boolean test that tells us whether or not changing a simple text item is
"cheap" or not.

Change-Id: Ifd878b43e0ee6ee808784dddd12e30a8863191d8
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>