Add some edge case handling
First and most importantly, don't let a benchmark run indefinitely. At
least on TX1, it seems that things can occasionally lock up (hard).
After running fine for ~1.5 days, my TX1 "hung" on a benchmark,
seemingly for GPU related reasons (at least, so I am informed in dmesg).

To handle this, we wait only so long before giving up, logging the
failure, and not attempting to start any other tests. We do this as a
precaution, as if we run into this condition, it likely means either
that someone is developing, or the harwdare needs a reset. Such hardware
reset is left to the harness running qmlbench.

In addition, we also log crashes a bit more noisily, on the offchance
that they are interesting to know about.

Change-Id: Id8d9eb50a29eb295669e743169a8d3e69c737776
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>