Clean up BenchmarkRunner a bit
After the process split, BenchmarkRunner no longer runs multiple
benchmarks, so a lot of the code here was dead - so we can remove it.
While we're at it, move the benchmarks to Options - as that's where they
really belong.

This removes the final tie from BenchmarkRunner to the host mode process,
which makes things a little tidier still.

Change-Id: I180fd61d1eb7267d68f06642d78383d8c0afc55c
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>