Add a flag to mark benchmarks as having a "boolean" result
Some of our benchmarks don't have a meaningful performance indicator (in
that they won't ever really improve or regress on the percentage
figure), but are still of interest in that they let us know if a
particular feature is totally broken or not.

This flag lets us easily mark such tests for ease of discovery when
looking into results / implementation of benchmarks.

Change-Id: Icda0981ba32d9abe84a835c0ca6ccb8a01be374d
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>