Replace delegates_script with delegates_item_script
delegates_script was a little flawed in that it set color as a prop, and
then set it again in the script handler. Simplify this by using item
(not rect) to bypass rendering.

This also tests a lot less stuff, and is now directly comparable with
delegates_item, which is a nice bonus.

Change-Id: I09a45e6dc1df2ea4cd116060a52a631bbfb9e92e
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>