Simple creation tests for a range of QtQuick types
The font-using ones are a bit of a pain in that we have to try to pick a
font that will likely always be there, the rest are quite
straightforward. View tests are not at all comprehensive, and may
deserve a set of dedicated tests of their own (with actual delegates,
and so on).

Change-Id: I6485e4cf62a3644519910786c32d91a490b657b4
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>