animations: Attempt to stabilize
Remove Grid and sizing stuff, use the bare minimum here. Create items
once, position them, and run animations repeatedly.

Also remove the PauseAnimation (which in itself is hard on predictable
work, which the benchmark kind of relies on) and Math.random (which
makes that even more hard to predict and not easily reproduced)

Change-Id: I3e74dca1ab8a302048575dcfb607cb84c2d2e822
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>