Make delegates_item_empty_jscreation & delegates_qobject comparable
In delegates_qobject, we now use a JS array to store items in before
deletion. This should make it a bit more heavy/slower.

In delegates_item_empty_jscreation, we remove the redundant parenting
(via 'root'), and just use the JS array. This should make it a bit
lighter/faster.

With these changes together, one can meaningfully get a bit of an idea
of the raw impact of allocating QQuickItem vs QObject.

Change-Id: I503a205760ff767b6ca8d42a2989755facaacb12
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>