qmlbench: Fix incorrect counting in frame-count shell
We should not count the warmup iterations in the overall number of
produced items. This seems wrong, and causes some confusion when
troubleshooting a broken test.

Note that a side effect of this change is that *all* results will now
drop by a value of 5 after this change is introduced.

Change-Id: Ib5e8a55411b663180f9d1a1a2a1df651bf0e4dc5
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>