Add internal random generator
In Qt 5.10 we introduce a new QRandomGenerator which is now
used by Math.random(), and there are signs that this is
significantly slower than qrand() which was used before. So
much so, that it is visible in many of the qmlbench benchmarks.
Note that this is a theory based on local runs and analysis of
the graphs and still needs to be proven.

Regardless, the qmlbench benchmarks are not intended to be
benchmarking Math.random(), it is just used to ensure the
positions tested are random. To eliminate that as a
possible cause for regressions, we instead introduce a
simple internal random function which just cycles through
a pregenerated array.

Note that this will reduce performance for all affected
tests on all branches of Qt, but it will hopefully reduce
the gap between Qt 5.9.0 and Qt 5.10.

Task-number: QTBUG-63922
Change-Id: Ib19c11323f38b23f9705e0d77a52a8a0f1157ac3
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>