Try to get a better random spread for tests with static count 20k
Some tests, where the static count was set to 20000 would collect
all items along a diagonal line. It seems as if the reason for this
was that QML would evaluate all Xs first and then all Ys, and since
the static count was an exact multiple of the size of the
pregenerated array, the multiplier for X and Y would be the same,
so they would gather along the diagonal.

By setting the size of the array we make this unlikely (tests
with static count 20002 will still encounter the issue, but we
have none of those.)

Change-Id: I2d1df3c17f270a185e73e53d4ec9e2b56ddc799d
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>