Default to Basic controls style if not already set
In Qt 6, the Default style was renamed to Basic, and the style that is
used by default now depends on which platform the application is run
on.

Previously the Default (now Basic) style was used by default.
So, restore the Basic style as the default. This ensures consistent
results should the default ever change again.

Task-number: QTBUG-88672
Change-Id: If96dd9cccd5fd05256221991bfacb4ae7bd3a1c1
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>