Fix disabling vsync for frame-count shell
The frame-count shell depends on disabling vsync to be able to
render as many frames as it can during 20 seconds. But the
logic was inverted in the condition that checked for the --shell
parameter, so the frame-count shell would never be detected and
the swap interval would never be set to 0.

This patch also updates the logic to take into consideration that
frame-count is the default now, so it will default to disabling
vsync.

Change-Id: I9d866a7fddfa46266ba1c8245b84e8389e00b154
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>